+++ data/user/colors/monokai.lua
@@ -19,6 +19,7 @@ style.line_highlight = {common.color "#36372f"}
 
 style.scrollbar  = {common.color "#49483E"}
 style.scrollbar2 = {common.color "#636254"}
+style.scrollbar3 = {common.color "#1b1c17"}
 
 style.syntax["normal"]   = {common.color "#F8F8F2"}
 style.syntax["symbol"]   = {common.color "#F8F8F2"}